reordering and cleanup
